-Ændret factory til constructor ved event- og changelog -Ændret id til Long fra Integer

+++ src/main/java/p3project/repositories/LogRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import p3project.classes.Eventlog;
 import p3project.classes.Changelog;
 
-public interface LogRepository extends JpaRepository<Eventlog, Integer>{
+public interface LogRepository extends JpaRepository<Eventlog, Long>{
 
 }
